backend/requirements.txt
fastapi
uvicorn
pandas-datareader
joblib
